tasks/main.yml: Handle nc apps properly in check mode
In check mode, the commands are not executed, so add 'check_mode: no'
to the task "lists the number of apps available in the instance.".
Since it may be run when testing installation (so occ command not
already available), ignore errors in test mode and use '|default("{}")'
to catch this "convert list to yaml.". The problem is that
"[NC apps] - installation." will fail since the dict would be empty
in this case, to run the loop only if not empty.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>